requirements.txt
python-dotenv
discord.py
requests
pinecone
langchain
langchain-pinecone==0.2.0
langchain-openai
langgraph
langchain-community
langgraph-checkpoint-sqlite